Unload message & confirmation on delete

+++ PhotoFox.Wpf.Ui.Mvvm/Commands/DeletePhotoCommand.cs
@@ -1,5 +1,7 @@
-﻿using PhotoFox.Storage.Blob;
+﻿using CommunityToolkit.Mvvm.Messaging;
+using PhotoFox.Storage.Blob;
 using PhotoFox.Storage.Table;
+using PhotoFox.Wpf.Ui.Mvvm.Messages;
 using PhotoFox.Wpf.Ui.Mvvm.ViewModels;
 using System;
 using System.Threading.Tasks;
@@ -15,14 +17,18 @@ namespace PhotoFox.Wpf.Ui.Mvvm.Commands
 
         private readonly IPhotoHashStorage photoHashStorage;
 
+        private readonly IMessenger messenger;
+
         public DeletePhotoCommand(
             IPhotoFileStorage photoFileStorage, 
             IPhotoMetadataStorage photoMetadataStorage,
-            IPhotoHashStorage photoHashStorage)
+            IPhotoHashStorage photoHashStorage,
+            IMessenger messenger)
         {
             this.photoFileStorage = photoFileStorage;
             this.photoMetadataStorage = photoMetadataStorage;
             this.photoHashStorage = photoHashStorage;
+            this.messenger = messenger;
         }
 
         public event EventHandler? CanExecuteChanged
@@ -44,11 +50,19 @@ namespace PhotoFox.Wpf.Ui.Mvvm.Commands
                 return;
             }
 
+            var msg = this.messenger.Send(new UserConfirmMessage("Are you sure you want to delete the selected photo?", "Warning"));
+            if (!msg.IsConfirmed)
+            {
+                return;
+            }
+
             await Task.WhenAll(
                 this.photoFileStorage.DeleteThumbnailAsync(selectedPhoto.RowKey),
                 this.photoFileStorage.DeletePhotoAsync(selectedPhoto.RowKey),
                 this.photoMetadataStorage.DeletePhotoAsync(selectedPhoto.PartitionKey, selectedPhoto.RowKey),
                 this.photoHashStorage.DeleteHashAsync(selectedPhoto.FileHash));
+
+            this.messenger.Send(new UnloadPhotoMessage(selectedPhoto));
         }
     }
 }
